refactor(worker): simplify job declaration #9
