+++ worker/Dockerfile
@@ -28,7 +28,5 @@ WORKDIR /app
 COPY package.json .
 RUN npm install
 COPY src src
-COPY launch.js .
 COPY tsconfig.json .
 RUN npm run build
-CMD [ "launch.js" ]
